feat: indicator 구현 및 ESM 개발 방식으로 변환
