+++ index.html
@@ -8,7 +8,7 @@
         <link rel="stylesheet" href="css/hero.css">
         <link rel="stylesheet" href="css/main.css">
         <link rel="stylesheet" href="css/footer.css">
-        <script src="main.js" defer></script>
+        <script type="module" src="js/app.js"></script>
     </head>
 
     <body>
